+++ NEWS.md
@@ -1,2 +1,8 @@
+## v0.17.0 2024-10-28
+
+- added --sf to ILS
+- passed to regular knot grid for bspline
+
 ## v0.16.0 2024-10-21
-- converted --wsd to use of gmresls
+
+- converted --wsd to use of gmresls in DLS